merged with 3.6.2

+++ bin/create_release.py
@@ -93,6 +93,7 @@ if diff_result:
     logging.warning("Index has non commited file/... The release will follow the last committed version.")
     answer = input('Do you want to continue anyway? (y/n)')
     if answer != 'y':
+        exit()
         
 release_date = date.fromtimestamp(time.time())
 for line in open(os.path.join(MG5DIR,'VERSION')):
@@ -118,9 +119,12 @@ auto_update = True
 
 # check that we are in the correct branch (note this file does not handle LTS)
 p = subprocess.Popen("git branch --show-current", stdout=subprocess.PIPE, shell=True)
-branch = p.stdout.read().decode().strip()
-if branch != '3.x':
-    print("cannot create tarball with auto-update outside of the main branch, detected branch (%s)" % branch)
+MG_branch = p.stdout.read().decode().strip()
+if MG_branch == 'LTS_2':
+    print("no auto-update as long as not the main version")
+    auto_update = False
+elif MG_branch not in  ['3.x', 'LTS']:
+    print("cannot create tarball with auto-update outside of the main branch, detected branch (%s)" % MG_branch)
     answer = input('Do you want to continue anyway? (y/n)')
     if answer != 'y':
         exit()
@@ -131,7 +135,6 @@ if branch != '3.x':
 if auto_update:
     p = subprocess.Popen(['git', 'tag', '-l', '-n','\'v%s\'' %version], stdout=subprocess.PIPE)
     old_tag = p.stdout.read().decode().strip()
-    print('vtag, old_tag')
     if old_tag:
         print("tag v%s is already existing. Those tags should be added by this script.")
         print("This will remove auto-update capabilities to this tarball")
@@ -139,7 +142,6 @@ if auto_update:
         if answer != 'y':
             exit()
         auto_update = False
-
     p = subprocess.Popen("git tag -l 'r*' ", stdout=subprocess.PIPE, shell=True)
     old_tag = p.stdout.read().decode().strip()
     print("detected release tag (bzr format)", old_tag)
@@ -176,22 +178,23 @@ elif auto_update:
 if auto_update:
     rev_nb_i = int(rev_nb)
     try:
-#        import ssl
-#        import urllib.request
-
-#        ctx = ssl.create_default_context()
-#        ctx.check_hostname = False
-#        ctx.verify_mode = ssl.CERT_NONE
-#        filetext = six.moves.urllib.request.urlopen('https://madgraph.mi.infn.it/mg5amc_build_nb', context=ctx)
-        filetext = six.moves.urllib.request.urlopen('http://madgraph.phys.ucl.ac.be/mg5amc3_build_nb')
+        if MG_branch == 'LTS':
+            import ssl
+            import urllib.request
+            ctx = ssl.create_default_context()
+            ctx.check_hostname = False
+            ctx.verify_mode = ssl.CERT_NONE
+            filetext = six.moves.urllib.request.urlopen('https://madgraph.mi.infn.it/mg5amc_build_nb', context=ctx)
+        elif MG_branch == '3.x':
+            filetext = six.moves.urllib.request.urlopen('http://madgraph.phys.ucl.ac.be/mg5amc3_build_nb')
         text = filetext.read().decode().split('\n')
-        print(text)
         web_version = int(text[0].strip())
         if text[1]:
             last_message = int(text[1].strip())
         else:
             last_message = 99
-    except (ValueError, IOError):
+    except (ValueError, IOError) as error:
+        print(error)
         logging.warning("WARNING: impossible to detect the version number on the web")
         answer = input('Do you want to continue anyway? (y/n)')
         if answer != 'y':
@@ -232,6 +235,7 @@ if auto_update:
 
 filepath = "MG5_aMC_v" + misc.get_pkg_info()['version'].replace(".", "_")
 filename = "MG5_aMC_v" + misc.get_pkg_info()['version'] + ".tar.gz"
+    
 if path.exists(filepath):
     logging.info("Removing existing directory " + filepath)
     shutil.rmtree(filepath)
@@ -256,8 +260,8 @@ for data in glob.glob(path.join(filepath, 'bin', '*')):
         else:
             os.rename(data, data.replace('compile.py','.compile.py'))
 
-os.remove(path.join(filepath, 'README.developer'))
-shutil.move(path.join(filepath, 'README.release'), path.join(filepath, 'README'))
+#os.remove(path.join(filepath, 'README.developer'))
+#shutil.move(path.join(filepath, 'README.release'), path.join(filepath, 'README'))
 
 
 # 1. Add information for the auto-update
@@ -268,7 +272,11 @@ if rev_nb and auto_update:
     fsock.write("last_message %s\n" % int(last_message))
     fsock.close()
     # tag handling
-    p = subprocess.call("git tag  'r%s' " % int(rev_nb), shell=True)
+    if MG_branch == 'LTS':
+        p = subprocess.call("git tag  'L%s' " % int(rev_nb), shell=True)
+    elif MG_branch == '3.x':
+        p = subprocess.call("git tag  'r%s' " % int(rev_nb), shell=True)
+if (rev_nb and auto_update) or MG_branch == "LTS_2":
     p = subprocess.call("git tag  'v%s' " % misc.get_pkg_info()['version'], shell=True)
     print('new tag added')
     answer = input('Do you want to push commit and tag? (y/n)')
@@ -279,9 +287,10 @@ if rev_nb and auto_update:
 # 1. Copy the .mg5_configuration_default.txt to it's default path
 shutil.copy(path.join(filepath, 'input','.mg5_configuration_default.txt'), 
             path.join(filepath, 'input','mg5_configuration.txt'))
-shutil.copy(path.join(filepath, 'input','.default_run_card_lo.dat'),
+if os.path.exists(path.join(filepath, 'input','.default_run_card_lo.dat')):
+    shutil.copy(path.join(filepath, 'input','.default_run_card_lo.dat'),
             path.join(filepath, 'input','default_run_card_lo.dat'))
-shutil.copy(path.join(filepath, 'input','.default_run_card_nlo.dat'),
+    shutil.copy(path.join(filepath, 'input','.default_run_card_nlo.dat'),
             path.join(filepath, 'input','default_run_card_nlo.dat'))
 shutil.copy(path.join(filepath, 'input','proc_card_default.dat'), 
             path.join(filepath, 'proc_card.dat'))
@@ -338,9 +347,11 @@ rm -rf download-temp;
 """ % filepath
 os.system(install_str)
 
-collier_link = "http://collier.hepforge.org/collier-latest.tar.gz" 
+sys.path.append(pjoin(root_path, '..'))
+from HEPToolsInstallers.HEPToolInstaller import _HepTools
+collier_link = _HepTools['collier']['tarball'][1] % _HepTools['collier']
+ninja_link = _HepTools['ninja']['tarball'][1] % _HepTools['ninja']
 misc.wget(collier_link, os.path.join(filepath, 'vendor', 'collier.tar.gz'))
-ninja_link = "https://bitbucket.org/peraro/ninja/downloads/ninja-latest.tar.gz"
 misc.wget(ninja_link, os.path.join(filepath, 'vendor', 'ninja.tar.gz'))
 
 # Add the tarball for SMWidth
@@ -371,7 +382,6 @@ except:
     logging.warning("Call to gpg to create signature file failed. " +\
                     "Please install and run\n" + \
                     "gpg --armor --sign --detach-sig " + filename)
-
 
 
 logging.info("Running tests on directory %s", filepath)
